Consolidate ReferClientContext and ReferServerContext
